+++ app/src/main/java/protect/card_locker/BarcodeImageWriterResultCallback.kt
@@ -1,0 +1,5 @@
+package protect.card_locker
+
+interface BarcodeImageWriterResultCallback {
+    fun onBarcodeImageWriterResult(success: Boolean)
+}