+++ app/src/main/java/protect/card_locker/async/CompatCallable.java
@@ -1,9 +1,0 @@
-package protect.card_locker.async;
-
-import java.util.concurrent.Callable;
-
-public interface CompatCallable<T> extends Callable<T> {
-    void onPostExecute(Object result);
-
-    void onPreExecute();
-}