+++ app/src/main/java/protect/card_locker/importexport/DataFormat.java
@@ -1,8 +1,0 @@
-package protect.card_locker.importexport;
-
-public enum DataFormat {
-    Catima,
-    Fidme,
-    Stocard,
-    VoucherVault;
-}