+++ app/src/main/java/protect/card_locker/importexport/DataFormat.kt
@@ -1,0 +1,7 @@
+package protect.card_locker.importexport
+
+enum class DataFormat {
+    Catima,
+    Fidme,
+    VoucherVault
+}